separate handlers for asyncio
