+++ pubnub/models/envelopes.py
@@ -1,0 +1,8 @@
+class AsyncioEnvelope:
+    def __init__(self, result, status):
+        self.result = result
+        self.status = status
+
+    @staticmethod
+    def is_error():
+        return False
