+++ tests/integrational/asyncio/test_signal.py
@@ -2,7 +2,8 @@ import pytest
 
 from pubnub.models.consumer.signal import PNSignalResult
 from pubnub.models.consumer.common import PNStatus
-from pubnub.pubnub_asyncio import PubNubAsyncio, AsyncioEnvelope
+from pubnub.pubnub_asyncio import PubNubAsyncio
+from pubnub.models.envelopes import AsyncioEnvelope
 from tests.integrational.vcr_helper import pn_vcr
 from tests.helper import pnconf_demo
